requirements.txt
streamlit>=1.40.0
google-cloud-bigquery>=3.25.0
google-auth>=2.35.0
db-dtypes>=1.3.0
pandas>=2.2.0
plotly>=5.24.0
pyyaml>=6.0.2
openpyxl>=3.1.0
